Change movement aread to static body 2D
We can set up collision bodies around the cirlce that blocks the player from moving.
This is nice because there's no maths/science to it.
We put it on a separate collision layer that only the player will look it.
The player will move around and naturally collide with the boundaries.